# app/routes/public_api.py
import re
from datetime import datetime

from flask import Blueprint, current_app, request

from ..services.award_book_service import AwardBookService
from ..utils.api_helpers import APIResponse, rate_limit, validate_isbn
from ..utils.error_handler import ErrorCategory, log_error
from ..utils.service_helpers import get_new_book_modules, get_service

public_api_bp = Blueprint('public_api', __name__, url_prefix='/api/public')

_award_service = AwardBookService()


@public_api_bp.route('/bestsellers')
@rate_limit(max_requests=60, window=60)
def get_all_bestsellers():
    try:
        limit = min(request.args.get('limit', 10, type=int), 50)
        book_service = get_service('book_service')
        if not book_service:
            return APIResponse.error('Service unavailable', 503, include_timestamp=True)
        categories = current_app.config.get('CATEGORIES', {})

        all_books = {}
        for cat_id, cat_name in categories.items():
            books = book_service.get_books_by_category(cat_id)
            all_books[cat_id] = {'category_name': cat_name, 'books': [book.to_dict() for book in books[:limit]]}

        return APIResponse.success(
            data={'categories': categories, 'books': all_books, 'last_updated': book_service.get_latest_cache_time()},
            include_timestamp=True,
        )

    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_all_bestsellers: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/bestsellers/<category>')
@rate_limit(max_requests=60, window=60)
def get_bestsellers_by_category(category: str):
    try:
        categories = current_app.config.get('CATEGORIES', {})
        if category not in categories:
            return APIResponse.error(
                f'Invalid category. Available categories: {list(categories.keys())}', 400, include_timestamp=True
            )

        limit = min(request.args.get('limit', 20, type=int), 50)
        book_service = get_service('book_service')
        if not book_service:
            return APIResponse.error('Service unavailable', 503, include_timestamp=True)
        books = book_service.get_books_by_category(category)

        return APIResponse.success(
            data={
                'category_id': category,
                'category_name': categories[category],
                'books': [book.to_dict() for book in books[:limit]],
                'total': len(books),
                'last_updated': book_service.get_latest_cache_time(),
            },
            include_timestamp=True,
        )

    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_bestsellers_by_category: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/bestsellers/search')
@rate_limit(max_requests=30, window=60)
def search_bestsellers():
    try:
        keyword = request.args.get('keyword', '').strip()

        if not keyword:
            return APIResponse.error('Keyword is required', 400, include_timestamp=True)
        if len(keyword) < 2:
            return APIResponse.error('Keyword must be at least 2 characters', 400, include_timestamp=True)
        if len(keyword) > 100:
            return APIResponse.error('Keyword must be at most 100 characters', 400, include_timestamp=True)
        if not re.match(r'^[\w\s\-\u4e00-\u9fff]+$', keyword):
            return APIResponse.error('Invalid keyword format', 400, include_timestamp=True)

        limit = min(request.args.get('limit', 20, type=int), 50)
        book_service = get_service('book_service')
        if not book_service:
            return APIResponse.error('Service unavailable', 503, include_timestamp=True)
        results = book_service.search_books(keyword)

        return APIResponse.success(
            data={'keyword': keyword, 'books': [book.to_dict() for book in results[:limit]], 'total': len(results)},
            include_timestamp=True,
        )

    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in search_bestsellers: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/awards')
@rate_limit(max_requests=60, window=60)
def get_all_awards():
    try:
        awards = _award_service.get_all_awards()

        book_counts = _award_service.get_book_counts_by_award(displayable_only=True)

        awards_data = []
        for award in awards:
            awards_data.append(
                {
                    'id': award.id,
                    'name': award.name,
                    'name_en': award.name_en,
                    'description': award.description,
                    'book_count': book_counts.get(award.id, 0),
                }
            )

        return APIResponse.success(data={'awards': awards_data, 'total': len(awards_data)}, include_timestamp=True)

    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_all_awards: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/awards/<award_name>')
@rate_limit(max_requests=60, window=60)
def get_award_books(award_name: str):
    try:
        award = _award_service.get_award_by_name(award_name)
        if not award:
            return APIResponse.error('Award not found', 404, include_timestamp=True)

        year = request.args.get('year', type=int)
        limit = min(request.args.get('limit', 20, type=int), 50)
        books, _ = _award_service.get_award_books(
            award_id=award.id, year=year, include_displayable_only=True, page=1, limit=limit
        )

        years = _award_service.get_distinct_years(award_id=award.id)

        return APIResponse.success(
            data={
                'award': {
                    'id': award.id,
                    'name': award.name,
                    'name_en': award.name_en,
                    'description': award.description,
                },
                'books': [book.to_dict() for book in books],
                'total': len(books),
                'years': years,
            },
            include_timestamp=True,
        )

    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_award_books: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/awards/<award_name>/<int:year>')
@rate_limit(max_requests=60, window=60)
def get_award_books_by_year(award_name: str, year: int):
    try:
        award = _award_service.get_award_by_name(award_name)
        if not award:
            return APIResponse.error('Award not found', 404, include_timestamp=True)

        books, _ = _award_service.get_award_books(
            award_id=award.id, year=year, include_displayable_only=True, page=1, limit=1000
        )

        if not books:
            return APIResponse.error(f'No books found for {award_name} in {year}', 404, include_timestamp=True)

        return APIResponse.success(
            data={
                'award': {'id': award.id, 'name': award.name, 'name_en': award.name_en},
                'year': year,
                'books': [book.to_dict() for book in books],
                'total': len(books),
            },
            include_timestamp=True,
        )

    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_award_books_by_year: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/book/<isbn>')
@rate_limit(max_requests=60, window=60)
def get_book_details(isbn: str):
    try:
        if not validate_isbn(isbn):
            return APIResponse.error('Invalid ISBN format', 400, include_timestamp=True)

        book_service = get_service('book_service')
        if not book_service:
            return APIResponse.error('Service unavailable', 503, include_timestamp=True)
        all_books = []
        for cat_id in current_app.config['CATEGORIES']:
            all_books.extend(book_service.get_books_by_category(cat_id))

        book = next((b for b in all_books if b.isbn13 == isbn), None)
        if book:
            return APIResponse.success(data={'book': book.to_dict(), 'source': 'bestseller'}, include_timestamp=True)

        award_book = _award_service.find_award_book_by_isbn(isbn)
        if award_book:
            return APIResponse.success(data={'book': award_book.to_dict(), 'source': 'award'}, include_timestamp=True)

        return APIResponse.error('Book not found', 404, include_timestamp=True)

    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_book_details: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/reports/weekly')
@rate_limit(max_requests=60, window=60)
def get_weekly_reports():
    try:
        from ..services.weekly_report_service import WeeklyReportService

        limit = min(request.args.get('limit', 10, type=int), 50)
        book_service = get_service('book_service')
        if not book_service:
            return APIResponse.error('Service unavailable', 503, include_timestamp=True)
        report_service = WeeklyReportService(book_service)
        reports = report_service.get_reports(limit)

        return APIResponse.success(
            data={'reports': [report.to_dict() for report in reports], 'total': len(reports)}, include_timestamp=True
        )

    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_weekly_reports: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/reports/weekly/latest')
@rate_limit(max_requests=60, window=60)
def get_latest_weekly_report():
    try:
        from ..services.weekly_report_service import WeeklyReportService

        book_service = get_service('book_service')
        if not book_service:
            return APIResponse.error('Service unavailable', 503, include_timestamp=True)
        report_service = WeeklyReportService(book_service)
        report = report_service.get_latest_report()

        if not report:
            return APIResponse.error('No report available', 404, include_timestamp=True)

        return APIResponse.success(data={'report': report.to_dict()}, include_timestamp=True)

    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_latest_weekly_report: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/reports/weekly/<date>')
@rate_limit(max_requests=60, window=60)
def get_weekly_report_by_date(date: str):
    try:
        from ..services.weekly_report_service import WeeklyReportService

        try:
            report_date = datetime.strptime(date, '%Y-%m-%d').date()
        except ValueError:
            return APIResponse.error('Invalid date format. Use YYYY-MM-DD', 400, include_timestamp=True)

        book_service = get_service('book_service')
        if not book_service:
            return APIResponse.error('Service unavailable', 503, include_timestamp=True)
        report_service = WeeklyReportService(book_service)
        report = report_service.get_report_by_date(report_date)

        if not report:
            return APIResponse.error('Report not found', 404, include_timestamp=True)

        return APIResponse.success(data={'report': report.to_dict()}, include_timestamp=True)

    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_weekly_report_by_date: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/new-books')
@rate_limit(max_requests=60, window=60)
def get_new_books():
    try:
        page = max(request.args.get('page', 1, type=int), 1)
        per_page = min(request.args.get('per_page', 20, type=int), 50)
        category = request.args.get('category')
        publisher_id = request.args.get('publisher_id', type=int)

        modules = get_new_book_modules()
        books, total = modules.query_service.get_new_books(
            publisher_id=publisher_id, category=category, page=page, per_page=per_page
        )

        return APIResponse.success(
            data={
                'books': [_serialize_new_book(b) for b in books],
                'total': total,
                'page': page,
                'per_page': per_page,
            },
            include_timestamp=True,
        )
    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_new_books: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/new-books/<publisher_name>')
@rate_limit(max_requests=60, window=60)
def get_new_books_by_publisher(publisher_name: str):
    try:
        page = max(request.args.get('page', 1, type=int), 1)
        per_page = min(request.args.get('per_page', 20, type=int), 50)

        modules = get_new_book_modules()
        publishers = modules.publisher_manager.get_publishers(active_only=True)
        publisher = next((p for p in publishers if p.name == publisher_name), None)
        if not publisher:
            return APIResponse.error('Publisher not found', 404, include_timestamp=True)

        books, total = modules.query_service.get_new_books(publisher_id=publisher.id, page=page, per_page=per_page)

        return APIResponse.success(
            data={
                'publisher': {'id': publisher.id, 'name': publisher.name},
                'books': [_serialize_new_book(b) for b in books],
                'total': total,
                'page': page,
                'per_page': per_page,
            },
            include_timestamp=True,
        )
    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_new_books_by_publisher: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


@public_api_bp.route('/recommendations')
@rate_limit(max_requests=60, window=60)
def get_recommendations():
    try:
        from ..services.recommendation_service import RecommendationService

        limit = min(request.args.get('limit', 10, type=int), 50)

        service = RecommendationService()
        result = service.get_smart_recommendations(limit=limit)

        return APIResponse.success(data=result, include_timestamp=True)
    except Exception as e:
        log_error(ErrorCategory.API_CALL, f'Error in get_recommendations: {e}')
        return APIResponse.error('Internal server error', 500, include_timestamp=True)


def _serialize_new_book(book) -> dict:
    data = {
        'id': book.id,
        'title': book.title,
        'author': book.author,
        'isbn13': book.isbn13,
        'category': book.category,
        'cover_url': book.cover_url,
        'publication_date': book.publication_date.isoformat() if book.publication_date else None,
    }
    if hasattr(book, 'title_zh') and book.title_zh:
        data['title_zh'] = book.title_zh
    if hasattr(book, 'publisher') and book.publisher:
        data['publisher'] = book.publisher.name if hasattr(book.publisher, 'name') else str(book.publisher)
    return data


@public_api_bp.route('/')
@rate_limit()
def api_info():
    return APIResponse.success(
        data={
            'name': 'BookRank Public API',
            'version': '1.2.0',
            'description': '提供畅销书排行榜和获奖图书数据的公开API',
            'endpoints': [
                {'path': '/api/public/bestsellers', 'method': 'GET', 'description': '获取所有分类畅销书'},
                {'path': '/api/public/bestsellers/<category>', 'method': 'GET', 'description': '获取指定分类畅销书'},
                {'path': '/api/public/bestsellers/search', 'method': 'GET', 'description': '搜索畅销书'},
                {'path': '/api/public/awards', 'method': 'GET', 'description': '获取所有奖项列表'},
                {'path': '/api/public/awards/<award_name>', 'method': 'GET', 'description': '获取指定奖项的获奖图书'},
                {
                    'path': '/api/public/awards/<award_name>/<year>',
                    'method': 'GET',
                    'description': '获取指定奖项和年份的获奖图书',
                },
                {'path': '/api/public/book/<isbn>', 'method': 'GET', 'description': '获取图书详细信息'},
                {'path': '/api/public/new-books', 'method': 'GET', 'description': '获取新书列表'},
                {'path': '/api/public/new-books/<publisher>', 'method': 'GET', 'description': '获取指定出版社新书'},
                {'path': '/api/public/recommendations', 'method': 'GET', 'description': '获取智能推荐'},
                {'path': '/api/public/reports/weekly', 'method': 'GET', 'description': '获取周报列表'},
                {'path': '/api/public/reports/weekly/latest', 'method': 'GET', 'description': '获取最新周报'},
                {'path': '/api/public/reports/weekly/<date>', 'method': 'GET', 'description': '根据日期获取周报'},
            ],
            'rate_limit': '60 requests per minute per IP',
            'documentation': 'https://github.com/gongyijie85/bookrank#api-documentation',
            'openapi': '/openapi.json',
        },
        include_timestamp=True,
    )